adding an iterm defaults recipe, refactoring rubymine prefs (didn't aram do this too?), adding a text mate link to mate

+++ cookbooks/pivotal_workstation/recipes/text_mate.rb
@@ -23,5 +23,6 @@ run_unless_marker_file_exists("text_mate_1_5_9") do
 
   execute "link textmate" do
     command "ln -s /Applications/TextMate.app/Contents/Resources/mate /usr/local/bin/mate"
+    not_if "test -e /usr/local/bin/mate"
   end
 end
